Changed float to double to prevent compilation error by buildbot.

+++ tracking/datcon/fpgaDATCON/findlets/src/DATCONSVDClusterLoaderAndPreparer.cc
@@ -65,9 +65,9 @@ void DATCONSVDClusterLoaderAndPreparer::apply(std::vector<std::pair<VxdID, std::
 
     if (cluster.isUCluster()) {
       nClusterPerLayer.at(layerNumber - 3)++;
-      const float rotangle = initialAngle[layerNumber - 3] + (ladderNumber - 1) * angleStep[layerNumber - 3];
-      const float cosRotAngle = cos(rotangle);
-      const float sinRotAngle = sin(rotangle);
+      const double rotangle = initialAngle[layerNumber - 3] + (ladderNumber - 1) * angleStep[layerNumber - 3];
+      const double cosRotAngle = cos(rotangle);
+      const double sinRotAngle = sin(rotangle);
       const int    sensorRadius = svdRadii[layerNumber - 3];
       const int    ytmp = localPosition + rPhiShiftsOfLayers[layerNumber - 3];
 
@@ -75,11 +75,11 @@ void DATCONSVDClusterLoaderAndPreparer::apply(std::vector<std::pair<VxdID, std::
       const long x = round(sensorRadius * cosRotAngle - ytmp * sinRotAngle); // x is in µm
       const long y = round(sensorRadius * sinRotAngle + ytmp * cosRotAngle); // y is in µm
       const long radiusSquared = x * x + y * y; // therefore radius is in µm as well, like all other length values
-//       const float radius = sensorRadius;
+//       const double radius = sensorRadius;
       // x or y divided by r^2 can create values very close to 0. To cope for this and still
       // have non-zero numbers after conversion and rounding, the conversion to int is made with 10^7
-      std::pair<long, long> pos2D = std::make_pair(convertToInt(2. * (float)x / (float)radiusSquared, 10),
-                                                   convertToInt(2. * (float)y / (float)radiusSquared, 10));
+      std::pair<long, long> pos2D = std::make_pair(convertToInt(2. * (double)x / (double)radiusSquared, 10),
+                                                   convertToInt(2. * (double)y / (double)radiusSquared, 10));
       uHits.emplace_back(std::make_pair(sensorID, pos2D));
     } else {
       nClusterPerLayer.at(4 + layerNumber - 3)++;
